pre-v0.6.2
MISRA-C:2004 Rules are validated

+++ boos/include/boos.system.thread.h
@@ -16,6 +16,6 @@
  *
  * @param millis a time to sleep in milliseconds.
  */  
-extern void threadSleep(int8 millis);
+extern void threadSleep(int16 millis);
 
 #endif /* BOOS_SYSTEM_THREAD_H_ */
